+++ Makefile
@@ -43,11 +43,9 @@ SRC += $(addprefix $(DIR_EXPAND), $(SRC_EXPAND))
 DIR_EXPAND:= parsing/expansion/
 SRC_EXPAND:=	expansion.c \
 				expansion_utils.c \
-				word_count_v2.c \
-				letter_count_v2.c \
+				letter_count.c \
+				word_count.c \
 				test/main_expand.c \
-#				letter_count.c \
-#				word_count.c \
 
 SRC +=  $(addprefix $(DIR_PARSE), $(SRC_PARSE))
 DIR_PARSE:= parsing/
